[main] Many improvements, still starting the application.
Hide borders and android buttons and toolbars that don't belong to my program.
FloorTables -> Which tables are present in the restaurant.
ShortTransactionList -> List of existing transactions that are open/closed/busy.
AskTransactionActivity -> Show floorTables and ShortTransactionList.
Round floortables.
Colour utils to brighten colours and make gradients.
ArcDrawable to make pie-chart circles.
Milk icons to show a glass full to empty.
Many xml improvements, have 8dp between all buttons and 0dp between bars.

+++ microfood/app/src/main/java/com/hha/service/DailyTransactionService.kt
@@ -401,11 +401,12 @@ class DailyTransactionService(channel: ManagedChannel) : BaseGrpcService<DailyTr
         }
     }
 
-    fun insertTransaction(tableName: String, customerId: Int, rfidKeyId: Int, transType: TransType): Int? = runBlocking {
+    fun insertTransaction(tableName: String, customerId: Int, rfidKeyId: Short, transType: TransType): Int? =
+        runBlocking {
         try {
             val request = InsertTransactionRequest.newBuilder()
                 .setTableName(tableName)
-                .setRfidKeyId(rfidKeyId)
+                .setRfidKeyId(rfidKeyId.toInt())
                 .setTransType(transType)
                 .setCustomerId(customerId)
                 .build()
